ajout de modifs sur les links du profile pro

+++ src/components/dashboard/professional/Reservation.tsx
@@ -4,7 +4,14 @@ import { Button } from "@/components/ui/button";
 import clsx from "clsx";
 import { CircleDot } from "lucide-react";
 import Image from "next/image";
-import { Dialog, DialogTrigger, DialogContent, DialogFooter, DialogHeader, DialogTitle } from "@/components/ui/dialog";
+import {
+  Dialog,
+  DialogTrigger,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
 
@@ -25,7 +32,7 @@ interface ReservationProps {
   firstName: string;
   lastName: string;
   reason?: string;
-  onDelete: (id: string) => void;
+  onDelete?: (id: string) => void;
 }
 
 export default function Reservation({
@@ -47,27 +54,36 @@ export default function Reservation({
   reason, // Raison d'annulation
   onDelete,
 }: ReservationProps) {
-  const displayAddress = typeof address === 'string' && address.trim() !== "" ? address : "Sur le lieu de travail";
+  const displayAddress =
+    typeof address === "string" && address.trim() !== ""
+      ? address
+      : "Sur le lieu de travail";
   const { user } = useUserContext();
   const [isDeleting, setIsDeleting] = useState(false);
   const [isDialogOpen, setIsDialogOpen] = useState(false);
   const [cancelReason, setCancelReason] = useState("");
-  const [isReasonDialogOpen, setIsReasonDialogOpen] = useState(false); 
+  const [isReasonDialogOpen, setIsReasonDialogOpen] = useState(false);
 
   const handleDelete = async () => {
     setIsDeleting(true);
     try {
       const response = await fetch(`/api/dashboardPro/delete/${id}`, {
-        method: 'DELETE',
+        method: "DELETE",
       });
       if (!response.ok) {
-        throw new Error('Erreur lors de la suppression de la réservation');
+        throw new Error("Erreur lors de la suppression de la réservation");
       }
       alert("Réservation supprimée avec succès !");
-      onDelete(id);
+      if (onDelete) {
+        onDelete(id);
+      }else{
+        console.log("onDelete props is missing, it's not passed to Reservation component");
+      }
     } catch (error) {
       console.error(error);
-      alert("Une erreur est survenue lors de la suppression de la réservation.");
+      alert(
+        "Une erreur est survenue lors de la suppression de la réservation."
+      );
     } finally {
       setIsDeleting(false);
     }
@@ -76,14 +92,14 @@ export default function Reservation({
   const handleCancelReservation = async () => {
     try {
       const response = await fetch(`/api/dashboardPro/cancel/${id}`, {
-        method: 'PUT',
+        method: "PUT",
         headers: {
-          'Content-Type': 'application/json',
+          "Content-Type": "application/json",
         },
         body: JSON.stringify({ reason: cancelReason }),
       });
       if (!response.ok) {
-        throw new Error('Erreur lors de l\'annulation de la réservation');
+        throw new Error("Erreur lors de l'annulation de la réservation");
       }
 
       const updatedReservation = await response.json();
@@ -101,7 +117,7 @@ export default function Reservation({
     try {
       const response = await fetch(`/api/dashboardPro/getReason/${id}`);
       if (!response.ok) {
-        throw new Error('Erreur lors de la récupération de la raison.');
+        throw new Error("Erreur lors de la récupération de la raison.");
       }
       const data = await response.json();
       setCancelReason(data.reason || "Aucune raison fournie.");
@@ -114,51 +130,65 @@ export default function Reservation({
   const handleHideReservation = async () => {
     setIsDeleting(true);
     try {
-      const response = await fetch(`/api/dashboardPro/updateReservationHidden/${id}`, {
-        method: 'PUT',
-      });
+      const response = await fetch(
+        `/api/dashboardPro/updateReservationHidden/${id}`,
+        {
+          method: "PUT",
+        }
+      );
       if (!response.ok) {
-        throw new Error('Erreur lors de la mise à jour de la réservation');
+        throw new Error("Erreur lors de la mise à jour de la réservation");
       }
       const data = await response.json();
       alert(data.message || "Réservation mise à jour avec succès !");
     } catch (error) {
       console.error(error);
-      alert("Une erreur est survenue lors de la mise à jour de la réservation.");
+      alert(
+        "Une erreur est survenue lors de la mise à jour de la réservation."
+      );
     } finally {
       setIsDeleting(false);
     }
   };
 
-  const formattedDate = new Date(date).toLocaleDateString('fr-FR', {
-    weekday: 'long', 
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
+  const formattedDate = new Date(date).toLocaleDateString("fr-FR", {
+    weekday: "long",
+    year: "numeric",
+    month: "long",
+    day: "numeric",
   });
 
-  const formattedTime = new Date(date).toLocaleTimeString('fr-FR', {
-    hour: '2-digit',
-    minute: '2-digit',
+  const formattedTime = new Date(date).toLocaleTimeString("fr-FR", {
+    hour: "2-digit",
+    minute: "2-digit",
   });
   return (
     <div className="bg-white flex md:justify-between justify-start md:flex-row flex-col gap-2 mb-4">
       <div className="shadow-md rounded-sm md:max-w-[16rem] w-full p-4 text-center border flex flex-col gap-3 justify-center items-center">
-        <div className={`${status === "boutique" ? "text-[#4C40ED] bg-[#F7F7FF]" : "text-[#FFA500] bg-[#FFF4E5]"
-          } py-2 px-3 rounded-md text-[.7rem]`}>
+        <div
+          className={`${
+            status === "boutique"
+              ? "text-[#4C40ED] bg-[#F7F7FF]"
+              : "text-[#FFA500] bg-[#FFF4E5]"
+          } py-2 px-3 rounded-md text-[.7rem]`}
+        >
           Client {typeClient === "boutique" ? "en boutique" : "flair"}
         </div>
         <Image
-          src={image || "/nail-salon.webp"}
+          src={image ? image : "/nail-salon.webp"}
           height={120}
           width={120}
           alt="client profile"
           className="rounded-full object-cover h-24 w-24"
         />
-        <h3>{firstName} {lastName}</h3>
+        <h3>
+          {firstName} {lastName}
+        </h3>
         <ul className="my-2">
           <li>
-            <span className="text-sm text-gray-500">Membre depuis dateCreation</span>
+            <span className="text-sm text-gray-500">
+              Membre depuis dateCreation
+            </span>
           </li>
           <li className="underline text-sm">{email}</li>
           <li className="underline text-sm">{phone}</li>
@@ -180,15 +210,32 @@ export default function Reservation({
               )}
             >
               <CircleDot className="h-4 w-4 inline-block mr-2" />
-              {status === "en-cours" ? "En cours" : status === "annule" ? "Annulée" : "Terminé"}
+              {status === "en-cours"
+                ? "En cours"
+                : status === "annule"
+                  ? "Annulée"
+                  : "Terminé"}
             </div>
             <ul className="text-sm text-gray-500 my-4">
-              <li className="mt-2"><strong>Service réservé :</strong> {service}</li>
-              <li className="mt-2"><strong>Date de réservation :</strong> {formattedDate} à partir de {formattedTime}</li>
-              <li className="mt-2"><strong>Durée :</strong> {dureeRDV}</li>
-              <li className="mt-2"><strong>Lieu :</strong> {displayAddress}</li>
-              <li className="mt-2"><strong>Tarifs :</strong> {price} €</li>
-              <li className="mt-2"><strong>Note client :</strong> {note}</li>
+              <li className="mt-2">
+                <strong>Service réservé :</strong> {service}
+              </li>
+              <li className="mt-2">
+                <strong>Date de réservation :</strong> {formattedDate} à partir
+                de {formattedTime}
+              </li>
+              <li className="mt-2">
+                <strong>Durée :</strong> {dureeRDV}
+              </li>
+              <li className="mt-2">
+                <strong>Lieu :</strong> {displayAddress}
+              </li>
+              <li className="mt-2">
+                <strong>Tarifs :</strong> {price} €
+              </li>
+              <li className="mt-2">
+                <strong>Note client :</strong> {note}
+              </li>
             </ul>
           </div>
           <div className="flex gap-4 items-end flex-wrap">
@@ -207,7 +254,9 @@ export default function Reservation({
                     placeholder="Expliquez la raison de l'annulation"
                   />
                   <DialogFooter>
-                    <Button onClick={() => setIsDialogOpen(false)}>Annuler</Button>
+                    <Button onClick={() => setIsDialogOpen(false)}>
+                      Annuler
+                    </Button>
                     <Button
                       variant="destructive"
                       onClick={handleCancelReservation}
@@ -221,9 +270,17 @@ export default function Reservation({
             )}
             {status === "annule" && (
               <>
-                <Dialog open={isReasonDialogOpen} onOpenChange={setIsReasonDialogOpen}>
+                <Dialog
+                  open={isReasonDialogOpen}
+                  onOpenChange={setIsReasonDialogOpen}
+                >
                   <DialogTrigger asChild>
-                    <Button variant="secondary" onClick={handleOpenReasonDialog}>Raison</Button>
+                    <Button
+                      variant="secondary"
+                      onClick={handleOpenReasonDialog}
+                    >
+                      Raison
+                    </Button>
                   </DialogTrigger>
                   <DialogContent>
                     <DialogHeader>
@@ -231,17 +288,27 @@ export default function Reservation({
                     </DialogHeader>
                     <p>{cancelReason}</p>
                     <DialogFooter>
-                      <Button onClick={() => setIsReasonDialogOpen(false)}>Fermer</Button>
+                      <Button onClick={() => setIsReasonDialogOpen(false)}>
+                        Fermer
+                      </Button>
                     </DialogFooter>
                   </DialogContent>
                 </Dialog>
-                <Button variant="destructive" onClick={handleHideReservation} disabled={isDeleting}>
+                <Button
+                  variant="destructive"
+                  onClick={handleHideReservation}
+                  disabled={isDeleting}
+                >
                   Supprimer
                 </Button>
               </>
             )}
             {status === "termine" && (
-              <Button variant="destructive" onClick={handleHideReservation} disabled={isDeleting}>
+              <Button
+                variant="destructive"
+                onClick={handleHideReservation}
+                disabled={isDeleting}
+              >
                 Supprimer
               </Button>
             )}
